commerce-engine/importexport/ep-importexport/src/test/java/com/elasticpath/importexport/importer/retrieval/impl/RetrievalMethodFactoryImplTest.java
/**
 * Copyright (c) Elastic Path Software Inc., 2013
 */
package com.elasticpath.importexport.importer.retrieval.impl;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertTrue;

import java.io.File;
import java.util.HashMap;
import java.util.Map;

import org.junit.Test;

import com.elasticpath.importexport.common.exception.ConfigurationException;
import com.elasticpath.importexport.common.types.TransportType;
import com.elasticpath.importexport.importer.configuration.RetrievalConfiguration;
import com.elasticpath.importexport.importer.retrieval.RetrievalMethod;

/**
 * RetrievalMethodFactoryImplTest. 
 */
public class RetrievalMethodFactoryImplTest {

	private final RetrievalMethodFactoryImpl retrievalMethodFactory = new RetrievalMethodFactoryImpl();
	
	/**
	 * Dummy Retrieval Method.
	 */
	private static class DummyRetrievalMethod extends AbstractRetrievalMethodImpl {
		@Override
		public File retrieve() {
			return null;
		}
	}
	
	/**
	 * testCreateRetrievalMethod.
	 * 
	 * @throws ConfigurationException if factory can't create.
	 */
	@Test
	public void testCreateRetrievalMethod() throws ConfigurationException {
		Map<TransportType, AbstractRetrievalMethodImpl> retrievalMethods = new HashMap<>();
		retrievalMethods.put(TransportType.FILE, new DummyRetrievalMethod());
		retrievalMethodFactory.setRetrievalMethods(retrievalMethods);
		assertEquals(retrievalMethods, retrievalMethodFactory.getRetrievalMethods());
		
		RetrievalConfiguration retrievalConfiguration = new RetrievalConfiguration();
		retrievalConfiguration.setMethod(TransportType.FILE);
		retrievalConfiguration.setSource("source");
		
		RetrievalMethod retrievalMethod = retrievalMethodFactory.createRetrievalMethod(retrievalConfiguration);
		
		assertNotNull(retrievalMethod);
		assertTrue(retrievalMethod instanceof DummyRetrievalMethod);
	}
	
	/**
	 * Tests Exception Throwing.
	 */
	@Test(expected = ConfigurationException.class)
	public void testCreateRetrievalMethodThrowsExeption() throws Exception {
		retrievalMethodFactory.setRetrievalMethods(new HashMap<>());
		
		RetrievalConfiguration retrievalConfiguration = new RetrievalConfiguration();
		retrievalConfiguration.setMethod(TransportType.FILE);
		retrievalConfiguration.setSource("source");
		
		
		retrievalMethodFactory.createRetrievalMethod(retrievalConfiguration);
	}
}
